All the stuff (#2)
* Placing towers

You can click on a tile and it will place a red square there. That red square is the instance of a class. This means we can put logic in it and we can call it in a loop every frame.

* bug fix

I didn't change the tilesize to the variable which is actaully based on screen size.

* tower selection

It's pretty bad, the selection options are kind of just added in there somewhat preprogrammed but reacting to sizing changes which guess isn't the worse.
It works on to the next thing.

* I did a thing